Renamed the car repository interface.
